+++ lib/airbrake-ruby/notice.rb
@@ -63,6 +63,8 @@ module Airbrake
     attr_reader :stash
 
     def initialize(config, exception, params = {})
+      params = deep_dup(params)
+
       @config = config
 
       @payload = {
@@ -238,6 +240,15 @@ module Airbrake
           "#{LOG_LABEL} #{exception.class}#to_airbrake failed:" \
           " #{attributes} must be a Hash"
         )
+      end
+    end
+
+    def deep_dup(params, level = 1)
+      level = 0 if level > 3
+      return params.dup if level == 0 || !params.respond_to?(:each_with_object)
+
+      params.each_with_object(params.dup) do |(key, value), hash|
+        hash[key] = deep_dup(value, level + 1)
       end
     end
   end